TheFiveinterFace/requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
openai==1.12.0
matplotlib==3.7.2
networkx==3.1
pydantic==2.5.2
asyncio==3.4.3
aiohttp==3.9.1
tiktoken==0.5.1 